requirements.txt
alabaster==0.7.12
altgraph==0.17
antiorm==1.2.1
argon2-cffi==20.1.0
astroid==2.4.2
attrs==20.1.0
autopep8==1.5.4
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.1.5
blinker==1.4
bluedot==1.3.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
eventlet==0.27.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Hashing==1.1
Flask-Mail==0.9.1
Flask-SocketIO==4.3.1
future==0.18.2
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
googlemaps==4.4.2
greenlet==0.4.16
httplib2==0.18.1
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
mysqlclient==2.0.1
nbconvert==5.6.1
nbformat==5.0.7
numpy==1.19.0
oauthlib==3.1.0
opencv-contrib-python==4.4.0.42
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyBluez==0.23
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
pylint==2.5.3
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.0
PyQt5-sip==12.8.1
PyQtWebEngine==5.15.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pyttsx3==2.71
pytz==2020.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.2
qrcode==6.1
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
rsa==4.6
Send2Trash==1.5.0
sense-hat==2.2.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0